Django-codebase/requirements.txt
APScheduler==3.6.1
asn1crypto==0.24.0
beautifulsoup4==4.8.0
bs4==0.0.1
CacheControl==0.12.5
cachetools==3.1.1
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
crypto==1.4.1
cryptography==2.7
Django==2.2.5
djangorestframework==3.10.3
firebase==3.0.1
firebase-admin==3.0.0
gcloud==0.18.3
google-api-core==1.14.2
google-api-python-client==1.7.11
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-cloud-core==1.0.3
google-cloud-firestore==1.4.0
google-cloud-storage==1.20.0
google-resumable-media==0.4.1
googleapis-common-protos==1.6.0
grpcio==1.24.1
httplib2==0.14.0
idna==2.8
jwcrypto==0.6.0
msgpack==0.6.2
Naked==0.1.31
oauth2client==4.1.3
protobuf==3.10.0rc1
pyasn1==0.4.7
pyasn1-modules==0.2.6
pycparser==2.19
python-jwt==3.2.4
pytz==2019.2
PyYAML==5.1.2
requests==2.22.0
rsa==4.0
shellescape==3.4.1
six==1.12.0
soupsieve==1.9.4
sqlparse==0.3.0
sseclient==0.0.24
tzlocal==2.0.0
uritemplate==3.0.0
urllib3==1.25.6
virtualenv==16.7.5
virtualenvwrapper-win==1.2.5
